nibbler:: add template DynamicLinkLibrary

+++ nibblerSources/class/gui/widget/WidgetOption.hpp
@@ -19,8 +19,8 @@ private:
 	bool &rNoise_;
 	bool &rMusique_;
 	std::string pathSound_;
-	int indexDisplayLibrary_;
-	int indexSoundLibrary_;
+	unsigned long indexDisplayLibrary_;
+	unsigned long indexSoundLibrary_;
 
 	bool musicManagemet_();
 	bool soundManagement_();
